+++ program/models/node_listing.py
@@ -5,7 +5,8 @@ from pypdf import PdfReader
 from langchain_core.prompts import ChatPromptTemplate
 from schemas.global_state import State
 from schemas.schema import KeywordDistribute, TitleOutput, BPOutput, DescriptionOutput
-from prompts.prompt_listing import keyword_prompt, title_prompt, bp_prompt, description_prompt, verification_prompt
+from prompts.prompt_listing import keyword_prompt, verification_prompt
+from utils.listing_func import generate_title, generate_bp, generate_description
 from langchain_openai import ChatOpenAI
 from utils.config_loader import config
 from dotenv import load_dotenv
@@ -55,252 +56,35 @@ def keyword_distribute(state: State):
         return {}
 
 # ====================================================================================================
-# Title 노드
-def generate_title(state: State):
-    
-    if not state['title_keyword']:
-        print('\n[Skipped] Title 작성용 키워드가 존재하지 않습니다.')
-        return {}
-    
-    print(f'\n--- Title 작성을 시작합니다... ---')
-    
-    try:
-        prompt = title_prompt.invoke(
-            {
-                'product_name': state['product_name'], 
-                'category': state['category'],
-                'product_information': state['product_information'], 
-                'title_keyword': state['title_keyword'],
-            }
-        )
-        structured_llm = llm.with_structured_output(TitleOutput)
-        res = structured_llm.invoke(prompt)
-        print(f'\n작성된 Title: 총 {len(res.title)}자')
-        return {'title': res.title}
-
-    except Exception as e:
-        print(f"\n[Error] Title 작성 중 에러가 발생했습니다: {e}")
-        return {}
-
-# ====================================================================================================
-# BP 노드
-def generate_bp(state: State):
-    
-    if not state['bp_keyword']:
-        print('\n[Skipped] Bullet Point 작성용 키워드가 존재하지 않습니다.')
-        return {}
-
-    print(f'\n--- Bullet Point 작성을 시작합니다... ---')
-    
-    try:
-        prompt = bp_prompt.invoke(
-            {
-                'product_name': state['product_name'], 
-                'category': state['category'],
-                'product_information': state['product_information'], 
-                'bp_keyword': state['bp_keyword'],
-            }
-        )
-        structured_llm = llm.with_structured_output(BPOutput)
-        res = structured_llm.invoke(prompt)
-        bp_length = []
-        for bp in res.bp:
-            bp_length.append(len(bp))    
-        print(f'\n작성된 Bullet Point: 각 {bp_length}자')
-        return {'bp': res.bp}
-    
-    except Exception as e:
-        print(f"\n[Error] Bullet Point 작성 중 에러가 발생했습니다: {e}")
-        return {}
-
-# ====================================================================================================
-# Description 노드
-def generate_description(state: State):
-    
-    if not state['description_keyword']:
-        print('\n[Skipped] Description 작성용 키워드가 존재하지 않습니다.')
-        return {}
-    
-    print(f'\n--- Description 작성을 시작합니다... ---')
-    
-    try:
-        prompt = description_prompt.invoke(
-            {
-                'bp_result': state['bp'],
-                'product_name': state['product_name'], 
-                'category': state['category'],
-                'product_information': state['product_information'], 
-                'description_keyword': state['description_keyword'],
-            }
-        )
-        structured_llm = llm.with_structured_output(DescriptionOutput)
-        res = structured_llm.invoke(prompt)
-        print(f'\n작성된 Description: 총 {len(res.description)}자')
-        return {'description': res.description}
-
-    except Exception as e:
-        print(f"\n[Error] Description 작성 중 에러가 발생했습니다: {e}")
-        return {}
-    
-
-# ====================================================================================================
-# Listing 한번에 생성하는 노드
+# 노드 작성
 def generate_listing(state: State):
-    
-    # Title 생성
-    if state['title_keyword']:
-        print(f'\n--- Title 작성을 시작합니다... ---')
-        
-        try:
-            prompt = title_prompt.invoke(
-                {
-                    'product_name': state['product_name'], 
-                    'category': state['category'],
-                    'product_information': state['product_information'], 
-                    'title_keyword': state['title_keyword'],
-                }
-            )
-            structured_llm = llm.with_structured_output(TitleOutput)
-            title = structured_llm.invoke(prompt)
-            print(f'\n작성된 Title: 총 {len(title.title)}자')
-
-        except Exception as e:
-            print(f"\n[Error] Title 작성 중 에러가 발생했습니다: {e}")
-            
-
-    else:    
-        print('\n[Skipped] Title 작성용 키워드가 존재하지 않습니다.')
-        return {}
-    
-    # BP 생성
-    if state['bp_keyword']:
-        print(f'\n--- Bullet Point 작성을 시작합니다... ---')
-    
-        try:
-            prompt = bp_prompt.invoke(
-                {
-                    'product_name': state['product_name'], 
-                    'category': state['category'],
-                    'product_information': state['product_information'], 
-                    'bp_keyword': state['bp_keyword'],
-                }
-            )
-            structured_llm = llm.with_structured_output(BPOutput)
-            bps = structured_llm.invoke(prompt)
-            bp_length = []
-            for bp in bps.bp:
-                bp_length.append(len(bp))    
-            print(f'\n작성된 Bullet Point: 각 {bp_length}자')
-        
-        except Exception as e:
-            print(f"\n[Error] Bullet Point 작성 중 에러가 발생했습니다: {e}")
-            return {}
-    
-    else:    
-        print('\n[Skipped] Bullet Point 작성용 키워드가 존재하지 않습니다.')
-        return {}
-        # Description 생성
-
-    if state['description_keyword']:
-        print(f'\n--- Description 작성을 시작합니다... ---')
-        
-        try:
-            prompt = description_prompt.invoke(
-                {
-                    'bp_result': state['bp'],
-                    'product_name': state['product_name'], 
-                    'category': state['category'],
-                    'product_information': state['product_information'], 
-                    'description_keyword': state['description_keyword'],
-                }
-            )
-            structured_llm = llm.with_structured_output(DescriptionOutput)
-            description = structured_llm.invoke(prompt)
-            print(f'\n작성된 Description: 총 {len(description.description)}자')
-
-        except Exception as e:
-            print(f"\n[Error] Description 작성 중 에러가 발생했습니다: {e}")
-            return {}
-        
-    else:
-        print('\n[Skipped] Description 작성용 키워드가 존재하지 않습니다.')
-        return {}
-    
-    return {'title': title.title, 'bp': bps.bp, 'description': description.description}
-
-# ====================================================================================================
-# Information Extract 노드
-INPUT_DATA_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "input_data")
-os.makedirs(INPUT_DATA_DIR, exist_ok=True)
-
-def information_extract(state: State) -> dict:
     """
-    정보 추출 노드 (information_extract)
-
-    목적:
-    - `program/input_data` 디렉토리 내의 모든 PDF 파일을 찾아 내용을 읽어옵니다.
-    - 읽어온 PDF 내용을 `State`의 `product_docs: List[Document]`에 저장합니다.
-    - LLM을 사용하여 `product_docs`의 내용을 요약하고, 핵심 제품 정보를 추출하여 `State`의 `product_information: str`에 저장합니다.
-
-    입력 (State에서 가져올 정보):
-    - (없음)
-
-    출력 (State에 업데이트할 정보):
-    - product_docs (List[Document]): PDF에서 읽어온 문서 객체들의 목록.
-    - product_information (str): 추출 및 요약된 핵심 제품 정보.
+    Title, Bullet Points, Description을 순차적으로 생성하기 위해
+    기존의 개별 생성 함수들을 호출하는 메인 함수.
     """
-    print("---정보 추출 노드 실행 중---")
-
-    product_docs: List[Document] = []
-    all_extracted_text = []
-
-    # PDF 파일 찾기 및 읽기
-    for filename in os.listdir(INPUT_DATA_DIR):
-        if filename.endswith(".pdf"):
-            filepath = os.path.join(INPUT_DATA_DIR, filename)
-            try:
-                reader = PdfReader(filepath)
-                text = ""
-                for page in reader.pages:
-                    text += page.extract_text() + "\n"
-                
-                product_docs.append(Document(page_content=text, metadata={"source": filename}))
-                all_extracted_text.append(text)
-                print(f'읽어온 PDF 파일 : {filename}')
-
-            except Exception as e:
-                print(f"PDF 파일 {filename} 읽기 오류: {e}")
-
-    # LLM을 사용하여 텍스트 요약
-    if all_extracted_text:
-        # 모든 PDF 내용을 하나의 문자열로 결합
-        full_text = "\n\n---\n\n".join(all_extracted_text)
-
-        # LLM 초기화 및 요약 체인 구성
-        llm = ChatOpenAI(model="gpt-4o", temperature=0)
+    print("\n--- 통합 리스팅 생성을 시작합니다... ---")
+    
+    # 1. Title 생성 함수 호출
+    title_update = generate_title(state)
+    
+    # 2. BP 생성 함수 호출
+    bp_update = generate_bp(state)
+    
+    # 3. Description 생성을 위해 state를 임시로 업데이트
+    # (generate_description 함수는 state에서 'bp' 결과를 필요로 함)
+    temp_state = state.copy()
+    temp_state.update(bp_update) # bp_update 딕셔너리 {'bp': [...]} 를 추가
         
-        summarization_template = ChatPromptTemplate.from_messages([
-            ("system", "You are an expert at summarizing product information. Please extract the key features, specifications, and purpose of the product from the following text in a concise manner. The summary should be in Korean."),
-            ("human", "Product Text:\n```\n{product_text}\n```")
-        ])
-        
-        summarization_chain = summarization_template | llm
-
-        print("---PDF 내용 요약 중---")
-        # LLM 호출하여 요약 생성
-        response = summarization_chain.invoke({"product_text": full_text})
-        product_information = response.content
-        print("---요약 완료---")
-        print(f"요약된 제품 정보: {product_information}")
-
-    else:
-        product_information = "제품 정보를 찾을 수 없습니다."
-
-    # State 업데이트
-    return {
-        "product_docs": product_docs,
-        "product_information": product_information
-    }
+    description_update = generate_description(temp_state)
+    
+    # 4. 모든 결과를 취합하여 최종 업데이트 딕셔너리 생xw성
+    final_update = {}
+    final_update.update(title_update)
+    final_update.update(bp_update)
+    final_update.update(description_update)
+    
+    print("\n--- 통합 리스팅 생성 완료 ---")
+    return final_update
 
 # ====================================================================================================
 # Listing Verification 노드
