+++ program/models/node_preprocess.py
@@ -5,9 +5,9 @@ from dotenv import load_dotenv
 from langchain_openai import ChatOpenAI
 from langchain_core.output_parsers import StrOutputParser
 from langchain.output_parsers import StructuredOutputParser, ResponseSchema
-from prompts.prompt_preprocess import filter_prompt, relevance_prompt, select_prompt, summarization_prompt
+from prompts.prompt_preprocess import filter_prompt, relevance_prompt, select_prompt, summarization_prompt, select_count
 from schemas.global_state import State
-from utils.func import preprocess_keywords, scaler
+from utils.preprocess_func import clean_keyword_column, filter_by_llm, clean_cp_column, clean_sv_column, scaler_and_score
 from utils.config_loader import config
 from langchain_core.prompts import ChatPromptTemplate
 
@@ -16,66 +16,34 @@ load_dotenv()
 
 # ====================================================================================================
 # 키워드 정제
-def keyword_preprocess(state: State):
-    
+def preprocess_data(state: State):
     if not state['data']:
         print("\n[Skipped] 데이터가 없어 키워드 정제를 종료합니다.")
         return sys.exit(1)
     
     try:
-        print(f'\n--- 키워드 {len(state['data'])}개에 대해 정제를 시작합니다... ---')
-        
-        llm = ChatOpenAI(model=config['llm_keyword']['model'], temperature=float(config['llm_keyword']['temperature']))
-       
+        """
+        데이터프레임 전체에 대해 정제, 스케일링, 점수 계산을 순차적으로 수행합니다.
+        """
+        print("\n--- 데이터 정제 및 스케일링 시작... ---")
         df = pd.DataFrame(state["data"])
-            
-        response_schemas = [
-            ResponseSchema(
-                name="keywords",
-                description="조건을 적용한 키워드 리스트, 문자열 리스트 또는 dict 리스트 가능"
-            )
-        ]
-        parser = StructuredOutputParser.from_response_schemas(response_schemas)
-        
-        # 1. keyword 사전 필터링
-        series = df['keyword']
-        """영어 알파벳/숫자/특수문자만 허용"""
-        series = series[series.apply(lambda x: bool(re.fullmatch(r"[A-Za-z0-9 &'().-]+", str(x))))]
-        filtered_series = series.drop_duplicates().reset_index(drop=True)
-        
-        # 2. 프롬프트 생성
-        keyword_prompt = filter_prompt.format(
-            data=filtered_series.tolist(),
-            format_instructions=parser.get_format_instructions()
-        )
+        df = clean_keyword_column(df)
+        df = filter_by_llm(df)
+        df, sv_imputed_mask = clean_sv_column(df)
+        df, cp_imputed_mask = clean_cp_column(df)
 
-        # 3. LLM 호출
-        res = llm.invoke([{"role": "user", "content": keyword_prompt}])
-        structured = parser.parse(res.content)
+        df['is_imputed'] = sv_imputed_mask | cp_imputed_mask
 
-        # 4. 파싱 처리
-        raw_keywords = structured.get("keywords", [])
-        if len(raw_keywords) > 0 and isinstance(raw_keywords[0], dict):
-            cleaned_keywords = [k.get('keyword') for k in raw_keywords if 'keyword' in k]
-        else:
-            cleaned_keywords = raw_keywords
+        df = scaler_and_score(df)
 
-        # 5. 원본 DF 필터
-        cleaned_data = df[df['keyword'].isin(cleaned_keywords)].reset_index(drop=True)
-        
-        processed_df = scaler(cleaned_data)
-        
-        processed_df = processed_df.to_dict(orient='records')
-        
-        print(f'\n키워드 {len(processed_df)}개를 정제하였습니다')
-        
-        return {"data": processed_df}
+        df.drop(columns=['is_imputed'], inplace=True, errors='ignore')
+        processed_df = df.to_dict(orient='records')
 
+        print(f"\n최종 {len(processed_df)}개 키워드 정제 및 점수 계산 완료.")
+        return {'data': processed_df}
     except Exception as e:
         print(f"\n[Error] 키워드 정제 중 에러가 발생했습니다: {e}")
         return {}
-
-
 
 
 # ====================================================================================================
@@ -173,6 +141,7 @@ def select_keywords(state: State) -> Dict:
             print(f"\n--- {len(simplified_data)}개 후보 중 상위 키워드 선별을 요청합니다... --- (시도 {retries + 1}/{max_retries})")
             
             response_str = chain.invoke({
+                'select_count': select_count,
                 "data_list_str": json.dumps(simplified_data, ensure_ascii=False)
             })
